Python: don't cleanup on SIGINT if -nocleanup option has been provided

+++ lib/mrtrix3/app.py
@@ -1165,7 +1165,7 @@ def handler(signum, _frame):
   sys.stderr.write('\n' + EXEC_NAME + ': ' + ANSI.error + msg + ANSI.clear + '\n')
   if os.getcwd() != WORKING_DIR:
     os.chdir(WORKING_DIR)
-  if SCRATCH_DIR:
+  if SCRATCH_DIR and DO_CLEANUP:
     try:
       shutil.rmtree(SCRATCH_DIR)
     except OSError:
